pyproject.toml
[tool.poetry]
name = "bgdia700"
version = "0.1.0"
description = "Webapp streamlit d'analyse de données"
authors = ["lolomasterIA <laury.magne@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
